+++ packages/types/src/actions.ts
@@ -67,6 +67,8 @@ import {
     FEEDBACK_STATUS,
     FEEDBACK_RESPONSE,
     SEND_TELEMETRY,
+    SHARE_LINK_TELEMETRY,
+    OPEN_LINK_TELEMETRY,
     UPDATEBOOKMARK,
     GETBOOKMARKS
 } from './types';
@@ -188,6 +190,10 @@ export const fillShareLinks = (payload: {
     type: FILL_SHARE_LINKS,
     payload
 });
+
+export const shareLinkTelemetry = () => ({ type: SHARE_LINK_TELEMETRY });
+
+export const openLinkTelemetry = () => ({ type: OPEN_LINK_TELEMETRY });
 
 export const restoreState = (payload: AppState): RestoreState => ({
     type: RESTORE_STATE,