+++ packages/webapp/src/webview/ui/components/Header/NavigationButtons/NavigationButtons.tsx
@@ -84,12 +84,17 @@ export function ShareButton() {
     const id = 'callout-test-id';
     const [isCopiedVisible, setCopiedVisible] = useState(false);
     const toggleCopied = (): void => {
+        actions.shareLinkTelemetry();
         setCopiedVisible(!isCopiedVisible);
     };
     const treeId = useSelector<AppState, GuidedAnswerTreeId>((state) => state.activeGuidedAnswer!?.TREE_ID);
     const nodes = useSelector<AppState, GuidedAnswerNode[]>((state) => state.activeGuidedAnswerNode);
     const nodeIdPath = nodes.map((n) => n.NODE_ID);
     const shareNodeLinks = useSelector<AppState, ShareNodeLinks | null>((state) => state.activeNodeSharing);
+
+    const copyInstructions = i18next
+        .t('COPY_WITH_INSTRUCTIONS_TEXT')
+        .replace('{EXTENSION_LINK}', shareNodeLinks?.extensionLink ?? '');
 
     return (
         <>
@@ -154,7 +159,19 @@ export function ShareButton() {
                                 </div>
                                 <p className="sharable-link__footer">{i18next.t('COPIED_TO_CLIPBOARD_DESC')}</p>
                                 <hr className="sharable-link__divider"></hr>
-                                <a className="sharable-link__web-link" id="web-link" href={shareNodeLinks.webLink}>
+                                <CopyToClipboard text={copyInstructions} onCopy={toggleCopied}>
+                                    <p
+                                        className="sharable-link__copy-instructions"
+                                        id="copy-instructions"
+                                        title={copyInstructions}>
+                                        {i18next.t('COPY_WITH_INSTRUCTIONS')}
+                                    </p>
+                                </CopyToClipboard>
+                                <a
+                                    className="sharable-link__web-link"
+                                    id="web-link"
+                                    href={shareNodeLinks.webLink}
+                                    onClick={() => actions.openLinkTelemetry()}>
                                     <UIIcon iconName={UiIcons.Export} />
                                     {i18next.t('VIEW_ON_WEBSITE')}
                                 </a>